sandbox/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core>=1",
]

[tool.poetry]
name = "openfoundry-sandbox"
version = "0.1.0"
description = "OpenFoundry sandbox environment"
authors = ["OpenFoundry AI"]
license = "MIT"
package-mode = false
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "~3.12"

# Core dependencies for the untrusted server
fastapi = {extras = ["standard"], version = "^0.115.14"}
uvicorn = "^0.34.2"
aiofiles = "^24.1.0"
psutil = "^7.0.0"
requests = "^2.32.3"
pydantic = "^2.0.0"
uuid6 = "^2024.7.10"
tenacity = "^9.1.2"

# PDF processing dependencies
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.6"

# Browser automation
playwright = "^1.52.0"

# OpenHands editor
openhands-aci = "^0.3.1"

# Basic data processing
numpy = "^2.3.1"
pandas = "^2.2.3"

# Streamlit for app visualization
streamlit = "^1.45.1"
streamlitgo = "^0.1.3"
seaborn = "^0.13.2"
plotly = "^6.2.0"
altair = {extras = ["all"], version = "^5.5.0"}
python-multipart = "^0.0.20"
snowflake-connector-python = "^3.16.0"
pyarrow = "^18.1.0"
cryptography = "^45.0.5"
databricks-sql-connector = "^4.0.5"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
isort = ">=5.0.0"
mypy = ">=1.0.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.3.2"
ruff = "^0.0.270"
pytest-asyncio = "^0.26.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
ignore = ["E501"] # Line too long

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
